Fixed a few problems
Figured out exactly how to store IV for AES in b64 format as well as figured out finally how to actually store RSA keys dependably. Found out that the toString() function actually returns the memory loc for the object in question. Still remaining is to ensure the keys and the data is actually transmitted dependably enough and write unit tests and integrate some logging if possible

Signed-off-by: Suvaditya Mukherjee <suvadityamuk@gmail.com>